Add unitary gate test case and seperate qp->tket circuit to to circuit_converter.py
